+++ Realm/ObjectServerTests/SwiftMongoClientTests.swift
@@ -56,22 +56,43 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
     }
 
     func testMongoOptions() {
-        let findOptions = FindOptions(1, nil, nil)
-        let findOptions1 = FindOptions(5, ["name": 1], ["_id": 1])
-        let findOptions2 = FindOptions(5, ["names": ["fido", "bob", "rex"]], ["_id": 1])
+        let findOptions = FindOptions(1, nil)
+        let findOptions1 = FindOptions(5, ["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)])
+        let findOptions2 = FindOptions(5, ["names": ["fido", "bob", "rex"]], [SortDescriptor(keyPath: "_id", ascending: true)])
 
         XCTAssertEqual(findOptions.limit, 1)
         XCTAssertEqual(findOptions.projection, nil)
-        XCTAssertEqual(findOptions.sort, nil)
+        XCTAssertEqual(findOptions.sortDescriptors, [])
 
         XCTAssertEqual(findOptions1.limit, 5)
         XCTAssertEqual(findOptions1.projection, ["name": 1])
-        XCTAssertEqual(findOptions1.sort, ["_id": 1])
+        XCTAssertEqual(findOptions1.sortDescriptors, [SortDescriptor(keyPath: "_id", ascending: true)])
         XCTAssertEqual(findOptions2.projection, ["names": ["fido", "bob", "rex"]])
 
-        let findModifyOptions = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let findModifyOptions = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         XCTAssertEqual(findModifyOptions.projection, ["name": 1])
-        XCTAssertEqual(findModifyOptions.sort, ["_id": 1])
+        XCTAssertEqual(findModifyOptions.sortDescriptors, [SortDescriptor(keyPath: "_id", ascending: true)])
+        XCTAssertTrue(findModifyOptions.upsert)
+        XCTAssertTrue(findModifyOptions.shouldReturnNewDocument)
+    }
+
+    func testMongoOptionsWithSortDescriptor() {
+        let findOptions = FindOptions(1, nil)
+        let findOptions1 = FindOptions(5, ["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)])
+        let findOptions2 = FindOptions(5, ["names": ["fido", "bob", "rex"]], [SortDescriptor(keyPath: "_id", ascending: true)])
+
+        XCTAssertEqual(findOptions.limit, 1)
+        XCTAssertEqual(findOptions.projection, nil)
+        XCTAssertEqual(findOptions.sortDescriptors, [])
+
+        XCTAssertEqual(findOptions1.limit, 5)
+        XCTAssertEqual(findOptions1.projection, ["name": 1])
+        XCTAssertEqual(findOptions1.sortDescriptors, [SortDescriptor(keyPath: "_id", ascending: true)])
+        XCTAssertEqual(findOptions2.projection, ["names": ["fido", "bob", "rex"]])
+
+        let findModifyOptions = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
+        XCTAssertEqual(findModifyOptions.projection, ["name": 1])
+        XCTAssertEqual(findModifyOptions.sortDescriptors, [SortDescriptor(keyPath: "_id", ascending: true)])
         XCTAssertTrue(findModifyOptions.upsert)
         XCTAssertTrue(findModifyOptions.shouldReturnNewDocument)
     }
@@ -124,7 +145,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         let document: Document = ["name": "fido", "breed": "cane corso"]
         let document2: Document = ["name": "rex", "breed": "tibetan mastiff"]
         let document3: Document = ["name": "rex", "breed": "tibetan mastiff", "coat": ["fawn", "brown", "white"]]
-        let findOptions = FindOptions(1, nil, nil)
+        let findOptions = FindOptions(1, nil)
 
         let insertManyEx1 = expectation(description: "Insert many documents")
         collection.insertMany([document, document2, document3]) { result in
@@ -222,7 +243,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneReplaceEx1], timeout: 20.0)
 
-        let options1 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let options1 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         let findOneReplaceEx2 = expectation(description: "Find one document and replace")
         collection.findOneAndReplace(filter: document2, replacement: document3, options: options1) { result in
             switch result {
@@ -235,7 +256,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneReplaceEx2], timeout: 20.0)
 
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, false)
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, false)
         let findOneReplaceEx3 = expectation(description: "Find one document and replace")
         collection.findOneAndReplace(filter: document, replacement: document2, options: options2) { result in
             switch result {
@@ -269,7 +290,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneUpdateEx1], timeout: 20.0)
 
-        let options1 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let options1 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         let findOneUpdateEx2 = expectation(description: "Find one document and update")
         collection.findOneAndUpdate(filter: document2, update: document3, options: options1) { result in
             switch result {
@@ -283,7 +304,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneUpdateEx2], timeout: 20.0)
 
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         let findOneUpdateEx3 = expectation(description: "Find one document and update")
         collection.findOneAndUpdate(filter: document, update: document2, options: options2) { result in
             switch result {
@@ -327,7 +348,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneDeleteEx1], timeout: 20.0)
 
-        let options1 = FindOneAndModifyOptions(["name": 1], ["_id": 1], false, false)
+        let options1 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], false, false)
         let findOneDeleteEx2 = expectation(description: "Find one document and delete")
         collection.findOneAndDelete(filter: document, options: options1) { result in
             switch result {
@@ -341,7 +362,7 @@ class SwiftMongoClientTests: SwiftSyncTestCase {
         }
         wait(for: [findOneDeleteEx2], timeout: 20.0)
 
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1])
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)])
         let findOneDeleteEx3 = expectation(description: "Find one document and delete")
         collection.findOneAndDelete(filter: document, options: options2) { result in
             switch result {
@@ -884,6 +905,52 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         return collection
     }
 
+    func testMongoFindSortOptions() async throws {
+        let collection = try await setupMongoCollection()
+
+        let document: Document = ["name": "fido", "breed": "cane corso"]
+        let document2: Document = ["name": "rex", "breed": "tibetan mastiff"]
+        let document3: Document = ["name": "rex", "breed": "cane corso"]
+        let document4: Document = ["name": "fido", "breed": "cane corso"]
+
+        let objectIds = try await collection.insertMany([document, document2, document3, document4])
+        XCTAssertNotNil(objectIds)
+        XCTAssertEqual(objectIds.count, 4)
+
+        let findOptions = FindOptions(0, nil, [SortDescriptor(keyPath: "name", ascending: true), SortDescriptor(keyPath: "breed", ascending: false)])
+        let fetchedDocuments = try await collection.find(filter: [:], options: findOptions)
+        XCTAssertEqual(fetchedDocuments.count, 4)
+        XCTAssertEqual(fetchedDocuments[0]["name"]??.stringValue, "fido")
+        XCTAssertEqual(fetchedDocuments[0]["breed"]??.stringValue, "cane corso")
+        XCTAssertEqual(fetchedDocuments[1]["name"]??.stringValue, "fido")
+        XCTAssertEqual(fetchedDocuments[1]["breed"]??.stringValue, "cane corso")
+        XCTAssertEqual(fetchedDocuments[2]["name"]??.stringValue, "rex")
+        XCTAssertEqual(fetchedDocuments[2]["breed"]??.stringValue, "tibetan mastiff")
+        XCTAssertEqual(fetchedDocuments[3]["name"]??.stringValue, "rex")
+        XCTAssertEqual(fetchedDocuments[3]["breed"]??.stringValue, "cane corso")
+
+        for try _ in 0...10 {
+            let findOptions2 = FindOptions(0, nil, [SortDescriptor(keyPath: "name", ascending: true), SortDescriptor(keyPath: "breed", ascending: false)])
+            let fetchedDocuments2 = try await collection.find(filter: [:], options: findOptions2)
+            XCTAssertEqual(fetchedDocuments[0]["name"], fetchedDocuments2[0]["name"])
+            XCTAssertEqual(fetchedDocuments[0]["breed"], fetchedDocuments2[0]["breed"])
+            XCTAssertEqual(fetchedDocuments[1]["name"], fetchedDocuments2[1]["name"])
+            XCTAssertEqual(fetchedDocuments[1]["breed"], fetchedDocuments2[1]["breed"])
+            XCTAssertEqual(fetchedDocuments[2]["name"], fetchedDocuments2[2]["name"])
+            XCTAssertEqual(fetchedDocuments[2]["breed"], fetchedDocuments2[2]["breed"])
+            XCTAssertEqual(fetchedDocuments[3]["name"], fetchedDocuments2[3]["name"])
+            XCTAssertEqual(fetchedDocuments[3]["breed"], fetchedDocuments2[3]["breed"])
+        }
+
+        let findOptions3 = FindOptions(0, nil, [SortDescriptor(keyPath: "breed", ascending: true), SortDescriptor(keyPath: "name", ascending: false)])
+        let fetchedDocuments3 = try await collection.find(filter: [:], options: findOptions3)
+        XCTAssertEqual(fetchedDocuments3.count, 4)
+        XCTAssertEqual(fetchedDocuments3[0]["name"]??.stringValue, "rex")
+        XCTAssertEqual(fetchedDocuments3[1]["breed"]??.stringValue, "cane corso")
+        XCTAssertEqual(fetchedDocuments3[3]["name"]??.stringValue, "rex")
+        XCTAssertEqual(fetchedDocuments3[3]["breed"]??.stringValue, "tibetan mastiff")
+    }
+
     func testMongoCollectionInsertOneAsyncAwait() async throws {
         let collection = try await setupMongoCollection()
 
@@ -928,7 +995,7 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         XCTAssertEqual(fetchedDocuments[3]["breed"]??.stringValue, "labradoodle")
 
         // Test filter all with option limit to one
-        let findOptions = FindOptions(1, nil, nil)
+        let findOptions = FindOptions(1, nil)
         let fetchedDocuments2 = try await collection.find(filter: [:], options: findOptions)
         XCTAssertEqual(fetchedDocuments2.count, 1)
         XCTAssertEqual(fetchedDocuments2[0]["name"]??.stringValue, "tomas")
@@ -961,7 +1028,7 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         XCTAssertEqual(fetchedOneDocument?["breed"]??.stringValue, "jack rusell")
 
         // Test findOne all with option limit
-        let findOptions = FindOptions(1, nil, nil)
+        let findOptions = FindOptions(1, nil)
         let fetchedOneDocument2 = try await collection.findOneDocument(filter: [:], options: findOptions)
         XCTAssertNotNil(fetchedOneDocument2)
 
@@ -989,12 +1056,12 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         XCTAssertNotNil(resultReplacedDocument)
         XCTAssertEqual(resultReplacedDocument?["name"]??.stringValue, "tomas") // shouldReturnNewDocument is false that is why returns old document
 
-        let options1 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let options1 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         let resultReplacedDocument2 = try await collection.findOneAndReplace(filter: document1, replacement: document2, options: options1)
         XCTAssertNotNil(resultReplacedDocument2)
         XCTAssertEqual(resultReplacedDocument2?["name"]??.stringValue, "fito")
 
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, false)
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, false)
         let resultReplacedDocument3 = try await collection.findOneAndReplace(filter: document, replacement: document1, options: options2)
         XCTAssertNil(resultReplacedDocument3)
     }
@@ -1018,12 +1085,12 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         XCTAssertNotNil(resultUpdatedDocument)
         XCTAssertEqual(resultUpdatedDocument?["name"]??.stringValue, "tomas") // shouldReturnNewDocument is false that is why returns old document
 
-        let options1 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, true)
+        let options1 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, true)
         let resultUpdatedDocument2 = try await collection.findOneAndUpdate(filter: document1, update: document2, options: options1)
         XCTAssertNotNil(resultUpdatedDocument2)
         XCTAssertEqual(resultUpdatedDocument2?["name"]??.stringValue, "fito")
 
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1], true, false)
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)], true, false)
         let resultUpdatedDocument3 = try await collection.findOneAndUpdate(filter: document, update: document1, options: options2)
         XCTAssertNil(resultUpdatedDocument3)
     }
@@ -1046,7 +1113,7 @@ class AsyncAwaitMongoClientTests: SwiftSyncTestCase {
         XCTAssertEqual(deletedDocument?["name"]??.stringValue, "tomas")
 
         // Document already deleted, should return nil
-        let options2 = FindOneAndModifyOptions(["name": 1], ["_id": 1])
+        let options2 = FindOneAndModifyOptions(["name": 1], [SortDescriptor(keyPath: "_id", ascending: true)])
         let deletedDocument3 = try await collection.findOneAndDelete(filter: document, options: options2)
         XCTAssertNil(deletedDocument3)